+++ .github/workflows/coordinator.yml
@@ -112,8 +112,6 @@ jobs:
     - name: Test coordinator packages
       working-directory: 'coordinator'
       run: |
-        make libzkp
-        # go test -exec "env LD_LIBRARY_PATH=${PWD}/libzkp/lib" -v -race -gcflags="-l" -ldflags="-s=false" -coverpkg="scroll-tech/coordinator" -coverprofile=coverage.txt -covermode=atomic ./...
         go test -v -race -gcflags="-l" -ldflags="-s=false" -coverprofile=coverage.txt -covermode=atomic -tags mock_verifier ./...
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v3
