+++ coordinator/internal/logic/libzkp/.gitignore
