+++ coordinator/internal/logic/libzkp/e2e-test.sh
