+++ coordinator/internal/logic/libzkp/lib.go
